# `generator-go-sdk`: speed up go sdk generate using goroutine to fetch service detail