--- a/tools/sdk/services/data.go
+++ b/tools/sdk/services/data.go
@@ -2,68 +2,12 @@ package services
 
 import (
 	"github.com/hashicorp/pandora/tools/sdk/resourcemanager"
+	"sync"
 )
 
 // GetResourceManagerServices returns all of the Services supported by the Resource Manager endpoint
 func GetResourceManagerServices(client resourcemanager.Client) (*ResourceManagerServices, error) {
-	services, err := client.Services().Get()
-	if err != nil {
-		return nil, err
-	}
-
-	resourceManagerServices := make(map[string]ResourceManagerService, 0)
-	for name, service := range *services {
-		serviceDetails, err := client.ServiceDetails().Get(service)
-		if err != nil {
-			return nil, err
-		}
-
-		terraformDetails, err := client.Terraform().Get(*serviceDetails)
-		if err != nil {
-			return nil, err
-		}
-
-		serviceVersions := make(map[string]ServiceVersion, 0)
-		for versionNumber, versionDetails := range serviceDetails.Versions {
-			versionInfo, err := client.ServiceVersion().Get(versionDetails)
-			if err != nil {
-				return nil, err
-			}
-
-			resources := make(map[string]Resource)
-			for resourceName, resourceDetails := range versionInfo.Resources {
-				operations, err := client.ApiOperations().Get(resourceDetails)
-				if err != nil {
-					return nil, err
-				}
-
-				schema, err := client.ApiSchema().Get(resourceDetails)
-				if err != nil {
-					return nil, err
-				}
-
-				resources[resourceName] = Resource{
-					Operations: *operations,
-					Schema:     *schema,
-				}
-			}
-
-			serviceVersions[versionNumber] = ServiceVersion{
-				Details:   *versionInfo,
-				Resources: resources,
-			}
-		}
-
-		resourceManagerServices[name] = ResourceManagerService{
-			Details:              service,
-			TerraformPackageName: serviceDetails.TerraformPackageName,
-			Terraform:            *terraformDetails,
-			Versions:             serviceVersions,
-		}
-	}
-	return &ResourceManagerServices{
-		Services: resourceManagerServices,
-	}, nil
+	return GetResourceManagerServicesByName(client, nil)
 }
 
 // GetResourceManagerServicesByName returns the specified Services from the Data API
@@ -79,59 +23,72 @@ func GetResourceManagerServicesByName(client resourcemanager.Client, servicesToL
 	}
 
 	resourceManagerServices := make(map[string]ResourceManagerService, 0)
+	var wg sync.WaitGroup
+	var lock sync.Mutex
 	for name, service := range *services {
-		if _, ok := serviceNames[name]; !ok {
-			continue
+		if len(serviceNames) > 0 {
+			if _, ok := serviceNames[name]; !ok {
+				continue
+			}
 		}
 
-		serviceDetails, err := client.ServiceDetails().Get(service)
-		if err != nil {
-			return nil, err
-		}
-
-		terraformDetails, err := client.Terraform().Get(*serviceDetails)
-		if err != nil {
-			return nil, err
-		}
-
-		serviceVersions := make(map[string]ServiceVersion, 0)
-		for versionNumber, versionDetails := range serviceDetails.Versions {
-			versionInfo, err := client.ServiceVersion().Get(versionDetails)
+		wg.Add(1)
+		go func(name string, service resourcemanager.ServiceSummary) {
+			defer wg.Done()
+			serviceDetails, err := client.ServiceDetails().Get(service)
 			if err != nil {
-				return nil, err
+				return
 			}
 
-			resources := make(map[string]Resource)
-			for resourceName, resourceDetails := range versionInfo.Resources {
-				operations, err := client.ApiOperations().Get(resourceDetails)
+			terraformDetails, err := client.Terraform().Get(*serviceDetails)
+			if err != nil {
+				return
+			}
+
+			serviceVersions := make(map[string]ServiceVersion, 0)
+			for versionNumber, versionDetails := range serviceDetails.Versions {
+				versionInfo, err := client.ServiceVersion().Get(versionDetails)
 				if err != nil {
-					return nil, err
+					return
 				}
 
-				schema, err := client.ApiSchema().Get(resourceDetails)
-				if err != nil {
-					return nil, err
+				resources := make(map[string]Resource)
+				for resourceName, resourceDetails := range versionInfo.Resources {
+					operations, err := client.ApiOperations().Get(resourceDetails)
+					if err != nil {
+						return
+					}
+
+					schema, err := client.ApiSchema().Get(resourceDetails)
+					if err != nil {
+						return
+					}
+
+					resources[resourceName] = Resource{
+						Operations: *operations,
+						Schema:     *schema,
+					}
 				}
 
-				resources[resourceName] = Resource{
-					Operations: *operations,
-					Schema:     *schema,
+				serviceVersions[versionNumber] = ServiceVersion{
+					Details:   *versionInfo,
+					Resources: resources,
 				}
 			}
 
-			serviceVersions[versionNumber] = ServiceVersion{
-				Details:   *versionInfo,
-				Resources: resources,
+			item := ResourceManagerService{
+				Details:              service,
+				TerraformPackageName: serviceDetails.TerraformPackageName,
+				Terraform:            *terraformDetails,
+				Versions:             serviceVersions,
 			}
-		}
-
-		resourceManagerServices[name] = ResourceManagerService{
-			Details:              service,
-			TerraformPackageName: serviceDetails.TerraformPackageName,
-			Terraform:            *terraformDetails,
-			Versions:             serviceVersions,
-		}
+			lock.Lock()
+			resourceManagerServices[name] = item
+			lock.Unlock()
+		}(name, service)
 	}
+	wg.Wait()
+
 	return &ResourceManagerServices{
 		Services: resourceManagerServices,
 	}, nil