much of time wasted waiting for responses from data API. Using goroutine to fetch service detail will profit performance much.

test results in my machine(with 6C16G in WSL2):
```bash
# main branch
./generator-go-sdk -output-dir ~/tmp  56.20s user 13.33s system 8% cpu 12:55.49 total

# this PR
./generator-go-sdk -output-dir ~/tmp  61.28s user 13.53s system 21% cpu 5:44.78 total
```